web/requirements.txt
alembic==0.9.1
appdirs==1.4.3
asn1crypto==0.22.0
bleach==2.0.0
blinker==1.4
captcha==0.2.2
cffi==1.10.0
click==6.7
cryptography==1.8.1
enum34==1.1.6
Flask==0.12
Flask-Breadcrumbs==0.4.0
Flask-Login==0.4.0
Flask-Menu==0.5.1
flask-paginate==0.4.5
Flask-Principal==0.4.0
Flask-WTF==0.14.2
html5lib==0.999999999
idna==2.5
ipaddress==1.0.18
itsdangerous==0.24
Jinja2==2.9.5
Mako==1.0.6
MarkupSafe==1.0
ndg-httpsclient==0.4.2
olefile==0.44
packaging==16.8
Pillow==4.1.1
pusher==1.7.0
pyasn1==0.2.3
pycparser==2.17
PyMySQL==0.7.10
pyOpenSSL==17.0.0
pyparsing==2.2.0
python-dateutil==2.6.0
python-editor==1.0.3
requests==2.13.0
six==1.10.0
SQLAlchemy==1.1.6
urllib3==1.21.1
webencodings==0.5
Werkzeug==0.12
WTForms==2.1
